% renamed widgets.dart file to suggestions.dart file inside widgets folder
